requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
Django==6.0.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
markdown-it-py==4.0.0
mdurl==0.1.2
openai==2.16.0
openpyxl==3.1.5
pillow==12.1.0
prompt_toolkit==3.0.52
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyreadline3==3.5.4
qrcode==8.2
rich==13.9.4
shell_gpt==1.5.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.1
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
wcwidth==0.5.0
